Update patch set 3
Patch Set 3:

Android's elfutils are used in two optional tools:

1. Elftree, is present on host on full Android eng builds but not included in SDK tool sets.  Elftree does a static analysis of object files and so does not need the decoding of stacked registers provided by the arch-dependent part of elfutils.

2. Perf, is present on host & target on full Android eng builds but not included in SDK tool sets. Support for Perf on mips targets is currently missing, in three ways:
a. Elfutils from upstream Fedora is missing Mips reg decode support; this can be gotten from Debian version of Elfutils.
b. Perf itself is missing an arch/mips/util/dwarf-regs.c for Mips; it does have it for arm and x86.  This is also missing from linux upstream Perf.
c. Perf requires the Android kernel to be compiled with perf support configured on; this is currently not yet done for the kernel for Mips.

If the Mips part of elfutils were added now, there is apparently direct way to test it on Android until Perf is made functional on Mips.

Patch-set: 3