Update patch set 1
Patch Set 1:

damn, i thought i'd already removed the version number from the filenames. we should do that before we take another update otherwise -- as you can see -- the diffs are useless and the history is trashed. (we also shouldn't have to touch the callers every time we upgrade libelf.)

we should rename '0.153' to 'src' before doing this upgrade.

Patch-set: 1