Update patch set 1
Patch Set 1:

> damn, i thought i'd already removed the version number from the
 > filenames. we should do that before we take another update
 > otherwise -- as you can see -- the diffs are useless and the
 > history is trashed. (we also shouldn't have to touch the callers
 > every time we upgrade libelf.)
 > 
 > we should rename '0.153' to 'src' before doing this upgrade.

I took this version number approach to be consistent but it sounds sensible.  

There seems to be 4 projects referencing the 0.153 directory ( elfutils, linux-tools-perf, ltrace and elftree ). We can drop any LOCAL_C_INCLUDES references and LOCAL_EXPORT_C_INCLUDE_DIRS to the elfutils libraries too.

This leaves elftree making a direct reference to the AndroidFixup for the host. It appears elfutils host files are for linux only , the AndroidFixup for linux is empty. 

I'll Abandon this one for now and do some cleanup first, If thats okay?  

Also Am I correct in thinking that using the same ChangeId on commit messages is the correct way to group patches from different projects for Deckard to be happy?

Patch-set: 1
Reviewer: Trevor Drake <1037213@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0