Update patch set 1
Patch Set 1:

I think what you want to verify is that
  git diff 2284375..86ed7ff   (change in upstream)
and
  git diff 5e22dcc..322dc30   (change to AOSP)
are the same.

They are the same in this merge, but I don't think gerrit has an option to compare "diff SHA1 SHA2" vs "diff SHA3 SHA4".

It's easy to see the diff with immediate parent commit through gitiles in gerrit:

On this review page, you can see its SHAs:
'Commit'    322dc30....  (gitiles)

Click on the (gitiles) link to see its content with the following SHAs:
parent 5e22dcc... [diff]
parent 86ed7f7... [diff]

Click on any one parent's [diff] link to see the diff from that parent.

It would be nice to get gitiles to display diff of any given SHAs,
but I have not found a way to do that.

Patch-set: 1